fix: adapt admin interface to work with the project
